refactor: CardDeck 컴포넌트 분리
