+++ src/components/TopicCards/index.tsx
@@ -1,25 +1,9 @@
-import { useState, useEffect, useRef } from "react";
-import { motion, AnimatePresence } from "framer-motion";
+import { useState, useEffect } from "react";
+import { motion } from "framer-motion";
 import * as S from "./styled";
 import { Topic } from "../../types/topic";
-import Card from "../Card";
-import { COLORS } from "../../styles/color";
 import IconArrow from "@assets/icons/left-arrow.svg";
-
-const SWIPE_THRESHOLD = -100;
-const SCALE_FACTOR = 0.05;
-const VERTICAL_OFFSET = 20;
-const DURATION = 0.83;
-
-const getColor = (index: number) => {
-  const colors = [
-    COLORS["--Primary-blue-500"],
-    COLORS["--Primary-blue-300"],
-    COLORS["--Primary-blue-100"],
-  ];
-  return colors[index % colors.length];
-};
-
+import CardDeck from "../CardDeck";
 interface TopicCardsProps {
   topics: Topic[];
   onHasViewedAllCards?: (hasViewedAllCards: boolean) => void;
@@ -28,29 +12,23 @@ interface TopicCardsProps {
 function TopicCards({ topics, onHasViewedAllCards }: TopicCardsProps) {
   const [currentIndex, setCurrentIndex] = useState(0);
   const [isLastSlide, setIsLastSlide] = useState(false);
-  const isAfterFirstRender = useRef(false);
 
   // 현재 보여줄 카드들 (최대 3개)
   const visibleTopics = topics.slice(currentIndex, currentIndex + 3);
 
-  const handleSwipeLeft = () => {
-    if (currentIndex < topics.length - 1) {
-      setCurrentIndex((prev) => prev + 1);
+  const handleSwipe = (direction: "left" | "right") => {
+    if (direction === "left") {
+      if (currentIndex < topics.length - 1) {
+        setCurrentIndex((prev) => prev + 1);
+      } else {
+        setIsLastSlide(true);
+      }
     } else {
-      setIsLastSlide(true);
+      if (currentIndex > 0) {
+        setCurrentIndex((prev) => prev - 1);
+      }
     }
   };
-
-  const handleSwipeRight = () => {
-    if (currentIndex > 0) {
-      setCurrentIndex((prev) => prev - 1);
-    }
-  };
-
-  useEffect(() => {
-    // 첫 렌더링 이후에 ref 값을 변경
-    isAfterFirstRender.current = true;
-  }, []);
 
   useEffect(() => {
     onHasViewedAllCards?.(isLastSlide);
@@ -70,76 +48,15 @@ function TopicCards({ topics, onHasViewedAllCards }: TopicCardsProps) {
         </motion.div>
       ) : (
         <>
-          <S.DeckWrapper>
-            {/* exit animation을 적용하기 위해 AnimatePresence 사용 */}
-            <AnimatePresence mode="popLayout">
-              {visibleTopics.map((topic, index) => (
-                <S.CardMotionWrapper
-                  key={topic.id}
-                  style={{
-                    zIndex: topics.length - index,
-                  }}
-                  // 첫 렌더링 시에는 animation 적용 안함
-                  // 첫번째 카드에 한해서 이후 initial을 적용하는데 뒤로가기 버튼을 클릭했을 때
-                  // 화면 밖으로 사라졌던(exit animation) 효과를 반대로 적용하기 위함
-                  initial={
-                    index === 0 && isAfterFirstRender.current
-                      ? {
-                          x: -300,
-                          y: -index * VERTICAL_OFFSET,
-                        }
-                      : false
-                  }
-                  animate={{
-                    scaleX: 1 - index * SCALE_FACTOR, // 뒤에 쌓인 카드는 작아짐
-                    x: 0,
-                    y: -index * VERTICAL_OFFSET, // 뒤에 쌓인 카드는 위로 올라감
-                    opacity: 1, // 카드가 보이도록
-                  }}
-                  // 퇴장 애니메이션(요소가 DOM에서 사라질 때 적용)
-                  // 첫번째 카드에만 적용. 다른 카드에도 적용해버리면 뒤로가기 버튼을 눌렀을 때 세번째 카드가 DOM에서 벗어나면서 퇴장 애니메이션이 적용됨
-                  exit={
-                    index === 0
-                      ? {
-                          x: -500,
-                          y: 0,
-                          transition: {
-                            duration: DURATION,
-                          },
-                          zIndex: 10,
-                        }
-                      : undefined
-                  }
-                  transition={{
-                    type: "spring", // 스프링 물리 기반 애니메이션
-                    stiffness: 200, // 스프링의 강성(높을수록 더 빠르고 탄력적)
-                    damping: 20, // 스프링의 감쇠(낮을수록 더 많이 튀어오름)
-                  }}
-                  onDragEnd={(_, info) => {
-                    if (info.offset.x < SWIPE_THRESHOLD) {
-                      handleSwipeLeft();
-                    } else if (info.offset.x > Math.abs(SWIPE_THRESHOLD)) {
-                      handleSwipeRight();
-                    }
-                  }}
-                  drag="x" // x축 방향으로만 드래그
-                  dragConstraints={{ left: 0, right: 0 }} // 드래그 제한 범위
-                  dragElastic={0.7} // 드래그 반발력(1에 가까울수록 더 탄력적)
-                >
-                  <Card
-                    content={topic.content}
-                    situationName={topic.situationName}
-                    id={topic.id}
-                    $color={getColor(index)}
-                  />
-                </S.CardMotionWrapper>
-              ))}
-            </AnimatePresence>
-          </S.DeckWrapper>
+          <CardDeck
+            visibleTopics={visibleTopics}
+            currentIndex={currentIndex}
+            onSwipe={handleSwipe}
+          />
           <S.TopicNavigationControls>
             <S.LeftButtonWrapper>
               {currentIndex > 0 && (
-                <S.IconButton onClick={handleSwipeRight}>
+                <S.IconButton onClick={() => handleSwipe("right")}>
                   <S.Icon src={IconArrow} alt="이전 토픽으로 돌아가기" />
                 </S.IconButton>
               )}
@@ -148,7 +65,7 @@ function TopicCards({ topics, onHasViewedAllCards }: TopicCardsProps) {
               {currentIndex + 1}/{topics.length}
             </S.ProgressText>
             <S.RightButtonWrapper>
-              <S.IconButton onClick={handleSwipeLeft}>
+              <S.IconButton onClick={() => handleSwipe("left")}>
                 <S.IconOpposite src={IconArrow} alt="다음 토픽으로 이동하기" />
               </S.IconButton>
             </S.RightButtonWrapper>
